# apps/api/app/modules/incidents/restapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RootLens API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/incidents": {
      "get": {
        "tags": ["Incidents"],
        "summary": "Get Incidents",
        "operationId": "get_incidents",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncidentResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Incidents"],
        "summary": "Create Incident",
        "operationId": "create_incident",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIncidentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{id}": {
      "get": {
        "tags": ["Incidents"],
        "summary": "Get Incident",
        "operationId": "get_incident",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Incident Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Incidents"],
        "summary": "Update Incident",
        "operationId": "update_incident",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIncidentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Incident Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ClientRequest": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" }
        }
      },
      "ClientResponse": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" }
        }
      },
      "ParticipantRequest": {
        "type": "object",
        "required": ["id", "name", "email"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string" }
        }
      },
      "CreateIncidentRequest": {
        "type": "object",
        "required": [
          "title",
          "severity",
          "description",
          "environment",
          "affected_clients"
        ],
        "properties": {
          "title": { "type": "string" },
          "severity": { "type": "string" },
          "description": { "type": "string" },
          "environment": { "type": "string" },
          "affected_clients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientRequest"
            }
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantRequest"
            },
            "default": []
          },
          "affects_all_clients": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "UpdateIncidentRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "severity": { "type": "string" },
          "description": { "type": "string" },
          "environment": { "type": "string" },
          "affected_clients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientRequest"
            }
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantRequest"
            }
          },
          "affects_all_clients": {
            "type": "boolean"
          }
        }
      },
      "EmployeeResponse": {
        "type": "object",
        "required": ["id", "display_name", "email", "role", "is_active"],
        "properties": {
          "id": { "type": "string" },
          "display_name": { "type": "string" },
          "email": { "type": "string" },
          "role": { "type": "string" },
          "is_active": { "type": "boolean" }
        }
      },
      "CommunicationChannelResponse": {
        "type": "object",
        "required": ["id", "display_name", "type", "link"],
        "properties": {
          "id": { "type": "string" },
          "display_name": { "type": "string" },
          "type": { "type": "string" },
          "link": { "type": "string" }
        }
      },
      "MitigationStepResponse": {
        "type": "object",
        "required": ["id", "order", "text", "owner"],
        "properties": {
          "id": { "type": "string" },
          "order": { "type": "integer" },
          "text": { "type": "string" },
          "owner": {
            "$ref": "#/components/schemas/EmployeeResponse"
          }
        }
      },
      "IncidentResponse": {
        "type": "object",
        "required": [
          "id",
          "public_id",
          "title",
          "description",
          "severity",
          "environment",
          "affected_clients",
          "affects_all_clients",
          "reported_date",
          "created_on",
          "modified_on",
          "participants",
          "status",
          "priority",
          "impact",
          "communication_channels",
          "mitigation_steps"
        ],
        "properties": {
          "id": { "type": "string" },
          "public_id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "severity": { "type": "string" },
          "environment": { "type": "string" },
          "affected_clients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientResponse"
            }
          },
          "affects_all_clients": {
            "type": "boolean"
          },
          "reported_date": {
            "type": "string",
            "format": "date-time"
          },
          "created_on": {
            "type": "string",
            "format": "date-time"
          },
          "modified_on": {
            "type": "string",
            "format": "date-time"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmployeeResponse"
            }
          },
          "status": { "type": "string" },
          "priority": { "type": "string" },
          "impact": { "type": "string" },
          "communication_channels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommunicationChannelResponse"
            }
          },
          "mitigation_steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MitigationStepResponse"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}